Temporarily remove graveseed
graveseed adds cognitive complexity in addition to not being ready.
It's what the author would want to use for writing IRC code,
but isn't necessarily what other developers might want to use.

Integrations will be added back in, however the goal is to
not require graveseed for much of the removed functionality.