Replace parser transforms with a Splitter
This is a farily broad change and more testing is definitely needed.
In particular, there may be edge cases around how it evaluates UTF-8
validity that are not properly tested for at this time.

This change gives a much nicer set of tools for parsing Bytes.
It eliminates a LOT of direct uses of from_unchecked and
opens up the possibility of simplifying Transform.