Improve ServerMsg memory footprint
Adds SharedSource; currently wraps Arc, will wrap ThinArc eventually.

Adds Source::owning_merged to minimize the memory footprint
of owning sources.

Uses SharedSource for ServerMsg::source.
This change moves 11 pointers (on platforms where usize == ptr)
out of ServerMsg in exchange for a net increase of up to
1 pointer and 2 atomic usizes (soon to be 1) per ServerMsg.
In practice it'll be less because Source is very sharable server-side.

Boxes the ServerMsg in the register handler error.