Forbid serde_derive versions with binary blobs
https://github.com/serde-rs/serde/issues/2538

Newer versions of serde_derive ship an unreproduced binary, likely
intended to improve proc macro performance by effectively
caching a compilation step. Nevertheless, this leads to supply chain
concerns; vinezombie largely limits its optional dependencies
to ones that are broadly trusted, but the binary (which has yet to be
reproduced or decompiled) cannot be reasonably trusted at this time.

This version restriction should hopefully be temporary, as either
the maintainer of serde will remove this anti-feature, or an adequate
fork of serde_derive will materialize that we can depend on instead.