Rework connection registration
Completely reworked nick generation AGAIN.

We now parse RPL_ISUPPORT and RPL_BOUNCE (both of them)!
The reg handler also now waits until the MOTD to finish.

The `MakeHandler` that creates a registration handler is now `Features`,
a struct that is used to extract information from an options struct.

Renamed `Register` to `Options` and trimmed it down a little.
Used by the `Features` that replace `ClientDefaults` and `BotDefaults`.

Changed default behaviors for connection registration.

Silly hashing for the Ping handler and elsewhere now uses SipHash.