Add and migrate to TagMaps
TagMaps are a way of mapping Tags to lazily-parsed values.
Additional information can be stored alongside the "natural" value for a tag.
This commit makes changes to FlatMap to support this new type.

Migrates ISUPPORT tokens to the new system,
and removes the serverinfo module (including ISupportParser).

Adds a module for common IRCv3 capabilities.
Also adds a public way to get SASL mechanisms out of the capability string.

Changes capabilities and ISUPPORT in Registration to use TagMaps.
This has particular benefits for capabilities, as now consumers of vinezombie
have access to info about unrequested capabilities.