Rework channel abstractions, add sync oneshot channels
Adds `oneshot` as essentially wrapped `Arc<OnceLock>`s,
but with a nicer interface and weak references on the sender side.
The senders can be cloned for usecases where one only cares about the first
operation to complete.

Adds `parker` for thread-parking, allowing the oneshot channels to block.
Very untested at the moment.

Changes the receiver type on Sender from `Arc<Self>` to `&mut Self`.
Also changes `ChannelSpec` and `*::make_channel` to match.
`Arc`s are not structural and oneshot-type channels really want to consume
the sender, which is quite clumsy when everything is an `Arc`.
Instead, `Sender` is impl'd for `Option`s containing oneshot senders.