Add Send bounds everywhere, rewrite oneshot
Client is now Send (but not Sync yet).

The new implementation of the oneshot channel has cloneable senders
and is Send for non-Sync values.
Non-consuming read-only access to the contained value via peek
is now behind an explicit T: Sync requirement.

Renames Receiver::recv_nonblocking to Receiver::recv_now to be more accurate.

Moves the modules in channel to their own files.