Make Register and defaults more generally-useful
Transposes the types in the return type for the password message.

The capabilities function in Register now returns by value, not reference.
The capability set is cloned in the handler creation method anyway,
and this adds a nicer place to modify the capability set as necessary.

Decouples the defaults functions from Options.
They can now be used for any type. Member access is handled by their callers.

Moves default_auth into a member on Options.

Prevents requesting echo-message by default.