Relicense under the EUPL
vinezombie does not (to the best of my knowledge) depend on any GPL'd code,
and as such I may choose to distribute my own contributions to vinezombie
(read: all of vinezombie) under a different license. This is me doing that.

I've deemed the EUPL to be more generally useful as it has network provisions
that aren't as misunderstood as the AGPL's, its copyleft provisions are more
concerned with whether the combined work is derivative than the exact
mechanism of combination, and it's broadly compatible with other copyleft
licenses (and I do think something that can constitute as large a part of
one's application as vinezombie should be copyleft).