Include ChannelSpec in Client
Oftentimes the channel spec isn't going to change frequently,
so it becomes fairly redundant as a parameter on Client::add.

This alters the Client creation functions to include the channel spec,
as well as adding a type parameter to Client for it.

Also removes new_client as it's basically pointless.