Always box handlers; use ControlFlow
Replaces SendCont with ControlFlow<Sent>,
and changes the return value of Handler::handle to a ControlFlow<()>.
This improves clarity, and ControlFlow has a (very limited) Try implementation.

MakeHandler now always boxes the handlers and no longer has an associated Handler type.
This improves flexibility and encapsulation, and we can afford the performance cost.