Add ClientCodec and ServerCodec
Adds a "tokio-codec" feature to implement the tokio_util
codec traits for these two types.

Copies the incremental parsing/writing functions on the
ircmsg types to these marker structs with the intent
of removing the ones on the types themselves.

Broadens some From impls for string pseudosubtypes.
This might affect inference, but shouldn't break anything in practice.